% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDataFilter.R
\name{calculateMAF}
\alias{calculateMAF}
\title{calculates minimum allele frequency (MAF)}
\usage{
calculateMAF(x)
}
\arguments{
\item{x}{A vector of marker allele data}
}
\value{
minimum allele frequency
}
\description{
A function for calculating minimum allele frequency. Genotype coding needs to be [0, 1, 2].
}
