% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToNumeric.R
\name{convertToNumeric}
\alias{convertToNumeric}
\title{convert to numeric}
\usage{
convertToNumeric(gData)
}
\arguments{
\item{x}{A data.frame or data.table of the dataset}
}
\value{
a data.frame or data.table of the dataset with all columns converted to numeric.
}
\description{
A function for converting an entire data.frame or data.table into numeric type.
}
