% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDataFilter.R
\name{filterGenoData}
\alias{filterGenoData}
\title{filters genotype dataset}
\usage{
filterGenoData(gData = genoDf, maf = 0.05, markerFilter = 0.6,
  indFilter = 0.8)
}
\arguments{
\item{gData}{A data.frame or data.table of the genotype dataset}

\item{maf}{Minimum allele frequency (MAF) used to filter out a marker, 0 to 1. By default, markers with MAF < 5 percent are removed.}

\item{markerFilter}{Fraction of missing values in a marker, 0 to 1. By default, markers with > 60 percent missing values are removed.}

\item{indFilter}{Fraction of missing values in an accession/genotype, 0 to 1. By default, accessions with > 80 percent missing marker values are removed.}
}
\value{
A data.frame of the cleaned genotype dataset.
}
\description{
A function for filtering genotype dataset based on minimum allele frequency, fraction of markers and accessions with missing values. It also filters out monomorphic markers.
}
