% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawGenoData.R
\docType{data}
\name{rawGenoData}
\alias{rawGenoData}
\title{Sample genotype dataset}
\format{
data.frame
}
\usage{
rawGenoData
}
\description{
A data.frame of genotype dataset for testing
}
\keyword{datasets}
