% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundAlleleDosage.R
\name{roundAlleleDosage}
\alias{roundAlleleDosage}
\title{rounds decimals}
\usage{
roundAlleleDosage(gData = gData, digits = 0)
}
\arguments{
\item{gData}{A data.frame or data.table of the dataset}

\item{digits}{an integer, decimal places to round. Defaults to 0.}
}
\value{
a data.frame or data.table of the genotype dataset with rounded values.
}
\description{
A function for rounding a dataset to a desired decimal place
}
