% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundAlleleDosage.R
\name{roundAlleleDosage}
\alias{roundAlleleDosage}
\title{rounds decimals}
\usage{
roundAlleleDosage(gData)
}
\arguments{
\item{x}{A data.frame or data.table of the genotype dataset}
}
\value{
a data.frame or data.table of the genotype dataset with rounded allele dosages.
}
\description{
A function for rounding decimals in genotype columns to create [0, 1, 2].
}
